refactor(code): consistent naming, wording, import constants
