+++ src/reducers/userReducer.js
@@ -1,15 +1,20 @@
+import {
+  USER_GENERATED,
+  USER_REGISTERED,
+} from '../actions/GetUserActions';
+
 const initialState = {
   data: null,
   registration: null,
 };
 
 const userReducer = (state = initialState, action) => {
-  if (action.type === 'USER_GENERATED') {
+  if (action.type === USER_GENERATED) {
     return Object.assign({}, state, {
       data: action.payload,
     });
   }
-  if (action.type === 'USER_REGISTERED') {
+  if (action.type === USER_REGISTERED) {
     return Object.assign({}, state, {
       registration: action.payload,
     });
